backend/event_service/requirements.txt
fastapi
uvicorn[standard]
redis
python-socketio
